Add support for stale data warnings
Some of Ryse's data sources are updated periodically, and might be
stale in ways that vary in severity by data source, and in ways that
might be unclear to a Blazer query user.

This change adds an optional stale data warning when running a query.
The business logic for the severity and content of the warning are
provided by the host application through a new configuration option,
which can be set in the host application via an initializer.